backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cramjam==2.7.0
Deprecated==1.2.14
distro==1.8.0
fastparquet==2023.10.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.6.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.2
openai==1.3.3
ordered-set==4.1.0
packaging==23.2
pandas==2.1.4
pyarrow==14.0.1
pydantic==2.5.1
pydantic_core==2.14.3
Pygments==2.17.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rich==13.7.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.16.0
